+++ Restaurants.Application/Restaurants/Commands/DeleteRestaurant/DeleteRestaurantCommandHandler.cs
@@ -1,5 +1,7 @@
 ﻿using AutoMapper;
-using MediatR;
+using MediatR;  
+using Restaurants.Domain.Interfaces;
+using Restaurants.Domain.Constants;
 using Restaurants.Domain.Entities;
 using Restaurants.Domain.Exceptions;
 using Restaurants.Domain.Repositories;
@@ -15,11 +17,13 @@ namespace Restaurants.Application.Restaurants.Commands.DeleteRestaurant
     {
         private readonly IRestaurantsRepository _restaurantsRepository;
         private readonly IMapper _mapper;
+        private readonly IRestaurantAuthorizationService _restaurantAuthorizationService;
 
-        public DeleteRestaurantCommandHandler(IRestaurantsRepository restaurantsRepository, IMapper mapper)
+        public DeleteRestaurantCommandHandler(IRestaurantsRepository restaurantsRepository, IMapper mapper, IRestaurantAuthorizationService restaurantAuthorizationService)
         {
             _restaurantsRepository = restaurantsRepository;
             _mapper = mapper;
+            _restaurantAuthorizationService = restaurantAuthorizationService;
         }
 
         public async Task Handle(DeleteRestaurantCommand request, CancellationToken cancellationToken)
@@ -29,6 +33,11 @@ namespace Restaurants.Application.Restaurants.Commands.DeleteRestaurant
             if (restaurant == null)
             {
                 throw new NotFoundException(nameof(Restaurant), request.Id.ToString());
+            }
+
+            if (!_restaurantAuthorizationService.Authorize(restaurant, ResourceOperation.Delete))
+            {
+                throw new ForbidException();
             }
 
             await _restaurantsRepository.DeleteRestauraunt(restaurant);
